refactor(app): restructure providers; implement routing and theming components for improved organization
