+++ src/app/providers/Providers.tsx
@@ -1,10 +1,10 @@
-import { BrowserRouter } from './router';
+import { AppRouter } from './router';
 import { CustomThemeProvider } from './theme';
 
 export const Providers = () => {
   return (
     <CustomThemeProvider>
-      <BrowserRouter />
+      <AppRouter />
     </CustomThemeProvider>
   );
 };
