--- src/app/providers/index.tsx
@@ -1,1 +1,0 @@
-export { Providers } from './providers';
