+++ src/app/providers/theme/ThemeProvider.types.ts
