+++ src/app/store/index.ts
@@ -1,1 +1,1 @@
-export { useAppStore } from './store';
+export { useAppStore } from './rootStore';
